requirements.txt
absl-py==2.1.0
Adafruit-Blinka==8.25.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-rfm9x==2.2.15
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.54.0
Adafruit-PureIO==1.1.11
appdirs==1.4.4
arandr==0.1.10
astroid==2.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
automationhat==0.2.0
backcall==0.2.0
backports.tarfile==1.1.1
beautifulsoup4==4.9.3
bleach==6.1.0
blinker==1.4
blinkt==0.1.2
build==1.2.1
buttonshim==0.0.2
cachetools==5.3.3
Cap1xxx==0.1.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.1
distro==1.5.0
docopt==0.6.2
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
executing==2.0.1
ExplorerHAT==0.4.2
fastjsonschema==2.20.0
Flask==1.1.2
flatbuffers==20181003210633
fourletterphat==0.1.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.62.2
gTTS==2.5.1
guizero==1.1.1
h5py==3.11.0
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
imutils==0.5.4
iniconfig==0.0.0
ipykernel==5.4.3
ipython==8.12.3
ipython_genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jaraco.context==5.3.0
jedi==0.18.0
Jinja2==3.1.4
joblib==0.17.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==0.0.0
libclang==18.1.1
logilab-common==1.8.1
lxml==4.6.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.3.4
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
microdotphat==0.2.1
mistune==3.0.2
ml-dtypes==0.4.0
mock==4.0.3
more-itertools==10.2.0
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.5
nudatus==0.0.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandocfilters==1.5.1
pantilthat==0.0.7
parso==0.8.1
pbr==5.5.0
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==0.13.0
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==5.8.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.12
pycairo==1.16.2
pycodestyle==2.6.0
pycups==2.0.1
pyflakes==2.2.0
pyftdi==0.55.0
pygame==1.9.6
Pygments==2.17.2
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pyLoRa==0.3.1
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproject_hooks==1.0.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
pytest==6.0.2
python-apt==2.2.1
python-dateutil==2.9.0.post0
python-prctl==1.7
python-vlc==3.0.20123
pyusb==1.2.1
pyzmq==26.0.3
QScintilla==2.11.6
qtconsole==5.0.2
QtPy==1.9.0
rainbowhat==0.1.0
referencing==0.35.1
regex==2020.11.13
reportlab==3.5.59
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.12.1
rich==13.7.1
roman==2.0.0
rpds-py==0.18.1
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
semver==2.10.2
Send2Trash==1.6.0b1
sense-emu==1.2
sense-hat==2.6.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
SpeechRecognition==3.10.3
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-cpu-aws==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thonny==4.0.1
tinycss2==1.3.0
toml==0.10.1
tomli==2.0.1
tornado==6.4.1
touchphat==0.0.1
tqdm==4.57.0
traitlets==5.14.3
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.11.0
uflash==1.2.4
unicornhathd==0.0.4
urllib3==2.2.1
v4l2-python3==0.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==3.0.2
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.16.0
xmltodict==0.13.0
yarg==0.1.9
zipp==3.18.1
